refactor: replace user authentication methods and token generation
